+++ api/app/dtos/organization/items/extended_dto.rb
@@ -4,7 +4,7 @@ class Organization::Items::ExtendedDto < OpenApiDto
   field "position", :integer
   field "name", :string
   field "description", :string, required: false
-  field "quantity", :integer
+  field "quantity", :decimal
   field "unit", :string
   field "unit_price_amount", :decimal
   field "item_group_id", :integer, required: false
